requirements.txt
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
codecov==2.0.22
colorama==0.4.3
coverage==5.1
Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
flake8==3.8.1
idna==2.9
Markdown==3.2.2
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
Pillow==7.1.2
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==5.4.2
pytest-cov==2.8.1
pytest-django==3.9.0
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.1.9
